+++ kgcnn/selection/__init__.py
@@ -1,3 +1,0 @@
-import logging
-
-logging.warning("Module 'kgcnn.selection' will be removed in future versions in favour of 'kgcnn.hyper'.")